refactor(casino): rename SyncBanner.jsx → SyncIcon.jsx
File exports SyncIcon (not a banner), so the name should match.
Update the import in study_casino.jsx accordingly.

BAZEL_TEST_INVOCATIONS=buildbuddy:84bbd8f6-d70c-49a8-8cd1-c8b075c3fb94
